% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoTube.R
\docType{data}
\name{ExamplePathways}
\alias{ExamplePathways}
\title{Example pathway database}
\format{
A list object with 30 vectors of gene symbols, for 30 pathways
}
\usage{
data(ExamplePathways)
}
\description{
A list object containing example gene sets from WikiPathways.
}
\keyword{datasets}
