% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoTube.R
\docType{data}
\name{ExampleResults}
\alias{ExampleResults}
\title{Example results from runLimmaAnalysis}
\format{
An MArrayLM object from limma
}
\usage{
data(ExampleResults)
}
\description{
Results of runLimmaAnalysis using the example data set GSE117751 
(in extdata).
}
\keyword{datasets}
