% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_mean.R
\name{gm_mean}
\alias{gm_mean}
\title{Calculate the geometric mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{Logical (default TRUE). Should NA values be ignored in this 
calculation? If FALSE, a vector containing NA values will return a 
geometric mean of NA.}
}
\value{
The geometric mean
}
\description{
Calculates the geometric mean of a numeric vector
}
\examples{
gm_mean(c(1, 3, 5))
}
