% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupByDistance.R
\name{groupByDistance}
\alias{groupByDistance}
\title{Cluster gene set analysis results}
\usage{
groupByDistance(mHG.res, l.edge, join.threshold = NULL,
  dist.method = "binary", returns = c("signif", "all"))
}
\arguments{
\item{mHG.res}{Results from pathway analysis for a single comparison, using mHG.genesets
or one member of mHG.geneset.multi. If an mHG.genesets.multi output, a single comparison
from the output list must be selected.}

\item{l.edge}{Leading edge result from mHG.to.LEdge or a single member of mHG.to.LEdge.multi.
If an mHG.to.LEdge.multi output, the matching comparison from mHG.res must be selected.}

\item{join.threshold}{The threshold distance to join gene sets. Gene sets with a distance
below this value will be joined to a single "cluster."}

\item{returns}{Either "signif" or "all". This argument defines whether only significantly
enriched gene sets are included in the output table, or if the full results are included.
Regardless of this selection, only significantly enriched gene sets (as defined in
mHG.to.LEdge or mHG.to.LEdge.multi) are clustered.}
}
\value{
A data frame including the mHG results, plus two additional columns for the clustering
results:
"Cluster" (The cluster that the gene set was assigned to. Gene sets in the same
cluster have a distance below the join.threshold) and "best" (Whether the gene set is the most enriched (by p-value) in a given cluster).
}
\description{
Groups the pathway analysis results (using mHG.genesets or mHG.genesets.multi) based on
the enriched gene sets' leading edges. If the calculated distance metric is lower than
the given threshold (i.e. the gene sets have highly overlapping leading edge genes),
these gene sets will be joined to a single gene set "cluster."
}
