% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.mHG.master.table.R
\name{make.mHG.master.table}
\alias{make.mHG.master.table}
\title{Make master table of all mHG results}
\usage{
make.mHG.master.table(genesetResults, leadingEdge, join.threshold = 0.5,
  filename = NULL)
}
\arguments{
\item{genesetResults}{Results from pathway analysis using mHG.geneset.multi.}

\item{leadingEdge}{Results from mHG.to.LEdge.multi.}

\item{join.threshold}{The threshold distance to join gene sets. Gene sets with a distance
below this value will be joined to a single "cluster."}

\item{filename}{File name for the output text file. If NULL (default), data will be returned
as an R data frame.}
}
\description{
This function clusters mHG results by leading edge similarity, and then combines to a
data frame or text file.
}
