% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDiffExprFile.R
\name{makeDiffExprFile}
\alias{makeDiffExprFile}
\title{Make differential expression results file.}
\usage{
makeDiffExprFile(limmaResults, base.group, filename = NULL,
  returns = c("all", "stats"))
}
\arguments{
\item{limmaResults}{Result from runLimmaAnalysis}

\item{base.group}{The same "base group" as in the Limma analysis.}

\item{filename}{The desired name for the output tab-delimited text file. If NULL (default)
the resulting table will be returned as an R data frame.}

\item{returns}{If "all" (default), will center the log-expression
data on median of base.group expression and include the expression data in the output. If
"stats", will only include the differential expression statistics.}
}
\description{
Make a data frame or text file containing coefficients, p-, and q-values from Limma
differential expression analysis. If returns == "all", will also center the log-expression
data on median of base.group expression and include the expression data in the output.
}
