% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNanoStringSetFromEset.R
\name{makeNanoStringSetFromEset}
\alias{makeNanoStringSetFromEset}
\title{Convert NanoString ExpressionSet to NanoStringSet}
\usage{
makeNanoStringSetFromEset(eset, designs = NULL)
}
\arguments{
\item{eset}{NanoString data ExpressionSet, from processNanostringData}

\item{designs}{Design matrix. If NULL, will look for "groups" column in 
pData(eset).}
}
\value{
A NanoStringSet for NanoStringDiff
}
\description{
Convert ExpressionSet from processNanoStringData to a NanoStringSet for use
in NanoStringDiff.
}
