% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHG.genesets.multi.R
\name{mHG.genesets.multi}
\alias{mHG.genesets.multi}
\title{Run all gene set analyses}
\usage{
mHG.genesets.multi(limmaResults, geneset.file, t_thresh = 2,
  numReq_overall = 10, numReq_inSet = 5)
}
\arguments{
\item{limmaResults}{Result from runLimmaAnalysis.}

\item{geneset.file}{Gene set file name, in .rds (list) or .gmt format. Gene names must be
in the same form as in the ranked.list.}

\item{t_thresh}{The threshold t-statistic for considering a gene positively or negatively.
For "negative" analysis, this threshold is set to -t_thresh.}

\item{numReq_overall}{The number of genes that must be differentially expressed above t_thresh
to conduct mHG analysis. If this is not satisfied, the comparison will be skipped.}

\item{numReq_inSet}{Number of genes required to conduct analysis on a given gene set (default = 1).
If fewer than this number of genes from the ranked list are included in a gene set, that
gene set will be skipped for this analysis.}
}
\value{
A list containing data frames with the mHG results for each comparison. Within each
comparison, both a "positive" and "negative" results data frame are provided, with mHG results
for genes most enriched in the comparison group vs. the base group (positive), and most enriched
in the base group vs. the comparison group (negative).
}
\description{
Runs minimum hypergeometric tests using the mHG library. This conducts mHG.genesets on every
possible comparison from the Limma analysis, in both the enriched (positive t-statistics) and
decreased (negative t-statistics) directions.
}
