% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHG.genesets.R
\name{mHG.genesets}
\alias{mHG.genesets}
\title{Run single gene set analysis.}
\usage{
mHG.genesets(ranked.list, geneset.file, numReq = 1,
  n_max = length(ranked.list))
}
\arguments{
\item{ranked.list}{A named vector, ranked in order of differential expression.}

\item{geneset.file}{Gene set file name, in .rds (list) or .gmt format. Gene names must be
in the same form as in the ranked.list.}

\item{numReq}{Number of genes required to conduct analysis on a given gene set (default = 1).
If fewer than this number of genes from the ranked list are included in a gene set, that
gene set will be skipped for this analysis.}

\item{n_max}{Maximum index for mHG test. By default, will allow the mHG test to include any
amount of genes from the list. This parameter can be adjusted to provide a maximum (e.g. the
first half of the list) length to be considered.}
}
\value{
A data frame containing the mHG results for each gene set.
}
\description{
Runs a minimum hypergeometric test on the ranked.list for the genes in every
gene set of the provided file, using the mHG library. The ranked.list is analyzed
from 1 to n_max. If analyses in both directions (beginning and end of list, or positively-
and negatively-enriched genes) are desired, mHG.genesets.multi can be used.
}
