% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHG.to.LEdge.multi.R
\name{mHG.to.LEdge.multi}
\alias{mHG.to.LEdge.multi}
\title{Process NanoString nCounter gene expression data.}
\usage{
mHG.to.LEdge.multi(mHG.res, limmaResults, t.thresh, q.thresh, geneset.file,
  dict = NULL, colnames_length = NULL)
}
\arguments{
\item{mHG.res}{Result from mHG.genesets.multi}

\item{limmaResults}{t matrix from runLimmaAnalysis result}

\item{t.thresh}{t-statistic threshold to include gene in leading edge}

\item{q.thresh}{q-value threshold to include pathway in analysis}

\item{geneset.file}{Same geneset file used in mHG.genesets.multi}
}
\description{
This function identifies leading edge genes from an mHG analysis result.
}
