% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mHG.to.LEdge.R
\name{mHG.to.LEdge}
\alias{mHG.to.LEdge}
\title{Conduct leading edge analysis for gene set analysis results}
\usage{
mHG.to.LEdge(mHG.res, t.stats, t.thresh, q.thresh, geneset.file, dict = NULL,
  colnames_length = NULL)
}
\arguments{
\item{mHG.res}{Result from mHG.genesets}

\item{t.stats}{t-stats from analysis}

\item{t.thresh}{t-statistic threshold to include gene in leading edge}

\item{q.thresh}{q-value threshold to include pathway in analysis}

\item{geneset.file}{Same geneset file used in mHG.genesets.multi}
}
\description{
This function identifies leading edge genes from an mHG analysis result. This is mostly
redundant with mHG.to.LEdge.multi.
}
