% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanostringPCA.R
\name{nanostringPCA}
\alias{nanostringPCA}
\title{Plot PCA}
\usage{
nanostringPCA(
  ns,
  pc1 = 1,
  pc2 = 2,
  interactive.plot = FALSE,
  exclude.zeros = TRUE
)
}
\arguments{
\item{ns}{Processed NanoString data}

\item{pc1}{Principal component to plot on x-axis (default 1)}

\item{pc2}{Principal component to plot on y-axis (default 2)}

\item{interactive.plot}{Plot using plotly? Default FALSE (in which case
ggplot2 is used)}

\item{exclude.zeros}{Exclude genes that are not detected in all samples
(default TRUE)}
}
\value{
A list containing:
\item{pca}{The PCA object}
\item{plt}{The PCA plot}
}
\description{
Conduct principal components analysis and plot the results.
}
