% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativeQC.R
\name{negativeQC}
\alias{negativeQC}
\title{Calculate negative control statistics}
\usage{
negativeQC(ns, interactive.plot = FALSE)
}
\arguments{
\item{ns}{NanoString data, processed by `processNanostringData` with
output.format set to 'list' and 'nSolver' normalization.}

\item{interactive.plot}{Generate an interactive plot using plotly? Only 
recommended for fewer than 20 samples (default FALSE)
\item{tab}{The table of negative control statistics, including the mean
& standard deviation of negative control genes, calculated background
threshold, and number of endogenous genes below that threshold}
\item{plt}{An object containing the negative control plots.}}

\item{numSD}{The number of standard deviations above the mean of negative
control gene expression to consider the "background" level (default 2).}
}
\description{
Provide a table the negative control statistics, and plot the counts of 
negative control genes in each sample.
}
