% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.housekeeping.R
\name{normalize.housekeeping}
\alias{normalize.housekeeping}
\title{Housekeeping gene normalization}
\usage{
normalize.housekeeping(dat, genes = NULL)
}
\arguments{
\item{dat}{NanoString data, including expression matrix and gene dictionary.}

\item{genes}{List of housekeeping genes to use for normalization. If NULL (default),
will use all genes marked as "Housekeeping" in codeset.}
}
\description{
Scale endogenous genes by the geometric mean of housekeeping genes. This should be conducted
after positive control normalization and background correction.
}
