% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.pos.controls.R
\name{normalize.pos.controls}
\alias{normalize.pos.controls}
\title{Positive control gene normalization}
\usage{
normalize.pos.controls(dat, logfile = "")
}
\arguments{
\item{dat}{NanoString data, including expression matrix and gene dictionary.}

\item{logfile}{Optional name of logfile to print warnings or errors.}
}
\description{
Scale genes by the geometric mean of positive control genes.
}
