% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_housekeeping.R
\name{normalize_housekeeping}
\alias{normalize_housekeeping}
\title{Housekeeping gene normalization}
\usage{
normalize_housekeeping(dat, genes = NULL, logfile = "")
}
\arguments{
\item{dat}{NanoString data, including expression matrix and gene dictionary.}

\item{genes}{List of housekeeping genes to use for normalization. If NULL (default),
will use all genes marked as "Housekeeping" in codeset.}
}
\description{
Scale endogenous genes by the geometric mean of housekeeping genes. This should be conducted
after positive control normalization and background correction.
}
