% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_pos_controls.R
\name{normalize_pos_controls}
\alias{normalize_pos_controls}
\title{Positive control gene normalization}
\usage{
normalize_pos_controls(dat, logfile = "")
}
\arguments{
\item{dat}{NanoString data, including expression matrix and gene dictionary.}

\item{logfile}{Optional name of logfile to print warnings or errors.}
}
\description{
Scale genes by the geometric mean of positive control genes.
}
