% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdiffToFGSEA.R
\name{nsdiffToFGSEA}
\alias{nsdiffToFGSEA}
\title{Run gene set enrichment analysis using DE results.}
\usage{
nsdiffToFGSEA(deResults, gene.sets, sourceDB = NULL, min.set = 1)
}
\arguments{
\item{gene.sets}{Gene set file name, in .rds (list), .gmt, or .tab format;
or a list object containing the gene sets. Gene names must be
in the same form as in the ranked.list.}

\item{sourceDB}{Source database to include, only if using a .tab-format 
geneset.file from CPDB.}

\item{min.set}{Number of genes required to conduct analysis on a given gene 
set (default = 1). If fewer than this number of genes from limmaResults are 
included in a gene set, that gene set will be skipped for this analysis.}

\item{limmaResults}{Result from runLimmaAnalysis.}
}
\value{
A list containing data frames with the fgsea results.
}
\description{
Use the fgsea library to run gene set enrichment analysis from the 
NanoStringDiff analysis results. Genes will be ranked by their log2 fold
changes.
}
