% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.stackedReport.from.grouped.mHG.R
\name{prep.stackedReport.from.grouped.mHG}
\alias{prep.stackedReport.from.grouped.mHG}
\title{Build leading edge report from clustered mHG results}
\usage{
prep.stackedReport.from.grouped.mHG(grouped.mHG, leadingEdge, limmaResults,
  base.group, filename = NULL)
}
\arguments{
\item{grouped.mHG}{The result from groupByDistance()}

\item{leadingEdge}{The same leading edge analysis used in groupByDistance()}

\item{limmaResults}{Result from runLimmaAnalysis()}

\item{base.group}{The same base group used for runLimmaAnalysis()}

\item{filename}{Optional name of output text file. If NULL (default), will return as data frame.}
}
\description{
This provides the genewise differential expression statistics and gene set membership for a
clustered mHG analysis.
}
