% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processNanostringData.R
\name{processNanostringData}
\alias{processNanostringData}
\title{Process NanoString nCounter gene expression data.}
\usage{
processNanostringData(fileDirs, bgType = c("threshold", "t.test"),
  bgThreshold = 3, bgProportion = 0.5, bgPVal = 0.001,
  housekeeping = NULL, skip.housekeeping = FALSE, includeQC = FALSE,
  sampIds = NULL, logfile = "")
}
\arguments{
\item{fileDirs}{file directory (or zip file) containing the .rcc files, or multiple directories in
a character vector.}

\item{bgType}{type of background correction to use: "threshold" sets a thresold for N standard deviations
above the mean of negative controls. "t.test" conducts a one-sided t test for each gene against all
negative controls.}

\item{bgThreshold}{If bgType=="threshold", number of sd's above the mean to set as threshold
for background correction.}

\item{bgProportion}{If bgType=="threshold", proportion of samples that a gene must be above threshold to
be included in analysis.}

\item{bgPVal}{If bgType=="t.test", p-value threshold to use for gene to be included in analysis.}

\item{housekeeping}{vector of genes (symbols or accession) to use for housekeeping correction. If NULL,
will use genes listed in .rcc files as "Housekeeping"}

\item{skip.housekeeping}{Skip housekeeping normalization?}

\item{includeQC}{Should we include the QC when reading in .rcc files? This can cause errors,
particularly when reading in files from multiple experiments.}

\item{sampIds}{a vector of sample identifiers, important if there are technical replicates.
Currently, this function averages technical replicates.}

\item{logfile}{a filename for the logfile (optional). If blank, will print warnings to screen.}
}
\value{
An rds file containing the raw and normalized counts, sample and qc info (from rcc files), and dictionary
}
\description{
This function reads in a zip file or folder containing multiple .rcc files, and then conducts positive
control normalization, background correction, and housekeeping normalization.
}
