% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gmt.R
\name{read.gmt}
\alias{read.gmt}
\title{Read .gmt file.}
\usage{
read.gmt(file)
}
\arguments{
\item{file}{The filename}
}
\value{
A list object, containing a character vector of genes for
each gene set.
}
\description{
Read in a .gmt gene set file. From the qusage library.
}
