% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cpdb_sourceDBs.R
\name{read_cpdb_sourceDBs}
\alias{read_cpdb_sourceDBs}
\title{Identify source databases from a .tab file}
\usage{
read_cpdb_sourceDBs(file)
}
\arguments{
\item{file}{The filename}
}
\value{
A table of the source databases, with the number of gene sets from
each one.
}
\description{
Read in a .tab file from the Consensus Pathway Database (CPDB), and
identify the source databases present.
}
