% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cpdb_tab.R
\name{read_cpdb_tab}
\alias{read_cpdb_tab}
\title{Read .tab file.}
\usage{
read_cpdb_tab(file, sourceDB = NULL)
}
\arguments{
\item{file}{The filename}

\item{sourceDB}{The source database to use. If NULL (default), retains
gene sets from all source databases}
}
\value{
A list object, containing a character vector of genes for
each gene set.
}
\description{
Read in a .tab file from the Consensus Pathway Database (CPDB)
}
