% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmt.R
\name{read_gmt}
\alias{read_gmt}
\title{Read .gmt file.}
\usage{
read_gmt(file)
}
\arguments{
\item{file}{The filename}
}
\value{
A list object, containing a character vector of genes for
each gene set.
}
\description{
Read in a .gmt gene set file. From the qusage library (Bolen C).
}
