% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_merge_rcc.R
\name{read_merge_rcc}
\alias{read_merge_rcc}
\title{Merge multiple .rcc files}
\usage{
read_merge_rcc(fileList, includeQC = FALSE, logfile = "")
}
\arguments{
\item{fileList}{a character vector of .rcc file names}

\item{includeQC}{include merged QC data (from the "Lane Attributes" part of file) in
the output? Default FALSE}

\item{logfile}{a filename for the logfile (optional). If blank, will print warnings to screen.}
}
\description{
Read in multiple .rcc files named in the fileList and merge the expression data.
}
