% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.background.R
\name{remove.background}
\alias{remove.background}
\title{Assess background expression}
\usage{
remove.background(dat, numSD, proportionReq, subtract = FALSE)
}
\arguments{
\item{dat}{Positive control-scaled NanoString data}

\item{numSD}{Number of standard deviations above mean of negative control genes to used as
background threshold for each sample: mean(negative_controls) + numSD * sd(negative_controls)}

\item{proportionReq}{Required proportion of sample expressions exceeding the sample background
threshold to include gene in further analysis}

\item{subtract}{Should calculated background levels be subtracted from reported expressions?
If TRUE, will subtract and then set negative values to zero (default FALSE)}
}
\description{
Compare endogenous gene expression data against negative control genes.
}
