% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_background.R
\name{remove_background}
\alias{remove_background}
\title{Assess background expression}
\usage{
remove_background(
  dat,
  mode = c("threshold", "t.test"),
  numSD,
  proportionReq,
  pval,
  subtract = FALSE
)
}
\arguments{
\item{dat}{Positive control-scaled NanoString data}

\item{mode}{Either "threshold" (default) or "t.test". If "threshold", requires proportionReq
of samples to have expression numSD standard deviations among the mean of negative control
genes. If "t.test", each gene will be compared with all negative control genes
in a one-sided two-sample t-test.}

\item{numSD}{Number of standard deviations above mean of negative control genes to used as
background threshold for each sample: mean(negative_controls) + numSD * sd(negative_controls).
Required if mode == "threshold" or subtract == TRUE}

\item{proportionReq}{Required proportion of sample expressions exceeding the sample background
threshold to include gene in further analysis. Required if mode == "threshold" or 
subtract == TRUE}

\item{pval}{p-value (from one-sided t-test) threshold to declare gene expression above
background expression level. Genes with p-values above this level are removed from
further analysis. Required if mode == "t.test"}

\item{subtract}{Should calculated background levels be subtracted from reported expressions?
If TRUE, will subtract mean+numSD*sd of the negative controls from the endogenous genes,
and then set negative values to zero (default FALSE)}
}
\description{
Compare endogenous gene expression data against negative control genes. and
remove data for genes that fail the comparison.
}
