% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLimmaAnalysis.R
\name{runLimmaAnalysis}
\alias{runLimmaAnalysis}
\title{Conduct differential expression analysis}
\usage{
runLimmaAnalysis(dat, groups = NULL, base.group = NULL, design = NULL)
}
\arguments{
\item{dat}{NanoString data ExpressionSet, from processNanostringData}

\item{groups}{character vector, in same order as the samples in dat. NULL
if already included in 'dat'}

\item{base.group}{the group against which other groups are compared (must
be one of the levels in 'groups'). Will use the first group if NULL.}

\item{design}{a design matrix for Limma analysis (default NULL, will do 
analysis based on provided 'group' data)}
}
\value{
The fit Limma object
}
\description{
Use Limma to conduct a simple differential expression analysis. All groups are
compared against the base.group, and empirical Bayes method is used to
identify significantly differentially expressed genes.
}
