% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untar_dirs.R
\name{untar_dirs}
\alias{untar_dirs}
\title{Untar}
\usage{
untar_dirs(fileDirs)
}
\arguments{
\item{fileDirs}{character list of tar files}
}
\value{
Names of now-untarred directories
}
\description{
Untars provided list of directories (analogous to unzip_dirs)
}
