% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip_dirs.R
\name{unzip_dirs}
\alias{unzip_dirs}
\title{Unzip}
\usage{
unzip_dirs(fileDirs)
}
\arguments{
\item{fileDirs}{character list of zip files}
}
\value{
Names of now-unzipped directories
}
\description{
Unzips provided list of directories
}
