% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmt.R
\name{read_gmt}
\alias{read_gmt}
\title{Read .gmt file.}
\usage{
read_gmt(file)
}
\arguments{
\item{file}{The filename}
}
\value{
A list object, containing a character vector of genes for
each gene set.
}
\description{
Read in a .gmt gene set file. From the qusage library (Bolen C).
}
\examples{
# Write a gmt file containing 2 gene sets.
gene_sets <- matrix(c("Set1", "Description", "GeneA", "GeneB",
                      "Set2", "Description", "GeneC", "GeneD"),
                    nrow = 2, ncol = 4, byrow = TRUE)
write.table(gene_sets, file = "foo.gmt", 
            row.names = FALSE, col.names = FALSE, sep = "\t", quote = FALSE)
            
# Read the gmt file
gene_set_list <- read_gmt("foo.gmt")

gene_set_list
}
